+++ app/Http/Controllers/RegisterTagihanController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\tagihan;
 use App\Models\register;
 use App\Models\register_tagihan;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Requests\Storeregister_tagihanRequest;
 use App\Http\Requests\Updateregister_tagihanRequest;
 
@@ -38,6 +39,9 @@ class RegisterTagihanController extends Controller
      */
     public function store(register $register, tagihan $tagihan)
     {
+        if (! Gate::allows('PPK', auth()->user()->id)&&! Gate::allows('Staf_PPK', auth()->user()->id)) {
+            abort(403);
+        }
         register_tagihan::create([
             'register_id'=>$register->id,
             'tagihan_id'=>$tagihan->id,
@@ -87,6 +91,9 @@ class RegisterTagihanController extends Controller
      */
     public function destroy(register $register, tagihan $tagihan)
     {
+        if (! Gate::allows('PPK', auth()->user()->id)&&! Gate::allows('Staf_PPK', auth()->user()->id)) {
+            abort(403);
+        }
         register_tagihan::where('tagihan_id', $tagihan->id)->where('register_id', $register->id)->delete();
         return redirect('/register/'. $register->id)->with('berhasil', 'Data Berhasil Di Hapus');      
     }
